+++ NEWS.md
@@ -1,6 +1,7 @@
 # textdata (development version)
 
 *  `lexicon_nrc_eil()` has been added to give access to the NRC Emotion Intensity Lexicon (aka Affect Intensity Lexicon) v0.5.
+*  `lexicon_nrc_vad()` has been added to give access to the The NRC Valence, Arousal, and Dominance Lexicon.
 
 # textdata 0.2.0